add bondset data model
